requirements.txt
ago==0.0.93
attrs==21.2.0
Automat==20.2.0
awscli==1.19.101
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.7.1
boto3==1.17.101
botocore==1.20.101
bs4==0.0.1
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
chromedriver-binary==92.0.4515.43.0
click==8.0.1
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cymem==1.31.2
cytoolz==0.9.0.1
dill==0.2.9
docutils==0.15.2
dotmap==1.3.23
elasticsearch==7.13.1
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
h2==3.2.0
hjson==3.0.2
hpack==3.0.0
hurry.filesize==0.9
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.5.0
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
langdetect==1.0.9
lxml==4.6.3
MarkupSafe==2.0.1
msgpack==0.6.2
msgpack-numpy==0.4.7.1
murmurhash==0.28.0
news-fetch==0.2.8
news-please==1.5.21
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.0
pandas==1.2.5
parsel==1.6.0
Pillow==8.2.0
plac==0.9.6
preshed==1.0.1
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyMySQL==1.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
PyYAML==5.4.1
queuelib==1.6.1
readability-lxml==0.8.1
regex==2017.4.5
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.4.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
Scrapy==2.5.0
selenium==3.141.0
sentencepiece==0.1.96
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
spacy==2.0.12
thinc==6.10.3
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
toolz==0.11.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.61.1
transformers==2.2.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.0
ujson==4.0.2
Unidecode==1.2.0
urllib3==1.26.6
w3lib==1.22.0
warcio==1.7.4
Werkzeug==2.0.1
wrapt==1.10.11
zipp==3.4.1
zope.interface==5.4.0
